BitArray, Conway's game of life

+++ src/events/mouse_scroll_event.hpp
@@ -9,8 +9,8 @@ class MouseScrollEvent : public Event {
 public:
     Point2I pos;
     ScrollDirection direction;
-    float horizontal;
-    float vertical;
+    f64 horizontal;
+    f64 vertical;
 
     MouseScrollEvent(const u64 timestamp, const i32 x, const i32 y,
                      const ScrollDirection direction,
